+++ Changelog.md
@@ -4,7 +4,50 @@ All notable changes to this project will be documented in this file
 The format is based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
-## [UNRELEASED]
+
+## [Unreleased]
+
+## 2.19.1
+### Added
+ - Added 'disable limiting response' option for select component with url as datasource.
+ 
+## 2.19.0
+### Added
+ - Form component to the form builder.
+ - JSON Logic on all places where custom javascript can be written.
+
+### Fixed
+ - Issues with the pages on the Wizard from showing the wrong pages or no pages at all.
+
+### Changed
+ - Upgraded ng-formio to 2.19.0
+ - Upgraded ng-dialog to 1.3.0
+ - Upgraded ng-tags-input to 3.2.0
+
+## 2.18.2
+### Added
+ - Add headers option to Select Urls.
+ 
+### Changed
+ - Upgrade ngFormio to 2.18.5
+
+### Changed
+ - Select resource components use Formio provider instead of $http.get to allow offline compatibility.
+
+## 2.18.1
+### Added
+ - A configuration for the resource references.
+
+### Changed
+ - Upgrade ngFormio to 2.18.1
+
+## 2.18.0
+### Fixed
+ - Custom components still didn't edit properly.
+ - Select url field missing due to passing ng-switch-when to field.
+ 
+### Changed
+ - Update ngFormio to 2.18.0
 
 ## 2.17.0
 ### Fixed